% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addclustvar.R
\name{addclustvar}
\alias{addclustvar}
\title{Adding constructed tariff classes}
\usage{
addclustvar(object, data)
}
\arguments{
\item{object}{GAM with frequency model}

\item{data}{data set with all x and y}
}
\value{
mutated data set contain a column of age groups that was set up using tariff classes
}
\description{
Adding constructed tariff classes for the outcome variable for frequency model
}
\details{
We first add the constructed tariff classes for the outcome variable to the data set, and sets the base level of the factor to the level with
the largest exposure.
}
\examples{
data(MTPL)
age_frequency <- fit_gam(data = MTPL, nclaims = nclaims, x = age_policyholder, exposure = exposure)
addclustvar(age_frequency,MTPL)
}
\author{
Seonghwa Min
}
