% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustplot.claimfreqency.R
\name{clustplot.claimfreqency}
\alias{clustplot.claimfreqency}
\title{plotting classes for claim frequency using Generalized Additive Models}
\usage{
clustplot.claimfreqency(object)
}
\arguments{
\item{object}{GAM with frequency model}
}
\value{
plot of age and predicted frequency divided with clusters
}
\description{
Plot showing classes for claim frequency (frequency = number of claims / exposure))
}
\details{
We want to set up classes and plot based on frequency using construct_tariff_classes for generalized additive models(GAM). The GAM model is an extension
of linear model to fit even the nonlinear data
}
\examples{
data(MTPL)
age_frequency <- fit_gam(data = MTPL, nclaims = nclaims, x = age_policyholder, exposure = exposure)
clustplot.claimfreqency(age_frequency)
}
\author{
Seonghwa Min
}
