% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustplot.claimseverity.R
\name{clustplot.claimseverity}
\alias{clustplot.claimseverity}
\title{plotting classes for claim severity using Generalized Additive Models}
\usage{
clustplot.claimseverity(object)
}
\arguments{
\item{object}{GAM with severity model}
}
\value{
plot of age and predicted severity divided with clusters
}
\description{
Plot showing classes for claim severity (loss, severity = severity / number of claims)
}
\details{
We want to set up classes and plot based on severity using construct_tariff_classes for generalized additive models(GAM). The GAM model is an extension
of linear model to fit even the nonlinear data. The plot will show age of the policy holder and the predicted severity.
}
\examples{
data(MTPL)
age_severity <- fit_gam(data = MTPL, nclaims = nclaims, x = age_policyholder, exposure = exposure, amount = amount, model = "severity")
clustplot.claimseverity(age_severity)
}
\author{
Seonghwa Min
}
