% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colin.R
\name{colin}
\alias{colin}
\title{Colinearity test}
\usage{
colin(data, y, crit = 10)
}
\arguments{
\item{data}{data set, preferrably data.frame}

\item{y}{outcome variable}

\item{crit}{criteria for variance inflation (default=10)}
}
\value{
list of boolean results regarding if VIF was over the criteria
}
\description{
Cmtplhecking for colinearity among variables
}
\details{
We want to check if any of the variables in the multivariate linear regression are correlated with each other. We will use VIF (variation inflation) to check the
co linearity using 10 as criteria (VIF>10)
}
\examples{
data(MTPL)
colin(MTPL,MTPL$nclaims,10)
}
\author{
Seonghwa Min
}
