% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmperf.R
\name{glmperf}
\alias{glmperf}
\title{Model performance (AIC, overdispersion, residuals)}
\usage{
glmperf(object1, object2)
}
\arguments{
\item{object1}{glm model 1 (typically null)}

\item{object2}{glam model 2 (typically alternative)}
}
\value{
plots and tables of the results from AIC, over dispersion and residuals
}
\description{
Model performance for glm
}
\details{
We want to check how the model performed compared to each other. Therefore we will print out a table with AIC, BIC and RMSE. Also, we will
check for overdispersion for both glm functions.Lastly, we will also check the normality assumption for the standardized residual by lookin at the qq plot
}
\examples{
data(MTPL)
model1<- glm(nclaims ~ freq_clusters, offset = log(exposure), family = "poisson", data = k)
model2 <- glm(nclaims ~ freq_clusters + age_policyholder, offset = log(exposure), family = "poisson", data = k)
glmperf(model1,model2)
}
\author{
Seonghwa Min
}
