% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmplotcomp.R
\name{glmplotcomp}
\alias{glmplotcomp}
\title{Plot comparing GLM models by tariff classes}
\usage{
glmplotcomp(object1, object2, model)
}
\arguments{
\item{object1}{glm model 1 (typically null)}

\item{object2}{glam model 2 (typically alternative)}

\item{model}{data set with tariff classes as a variable}
}
\value{
plot of rates by age group (tariff classes)
}
\description{
Plot by tariff classes comparing glm models
}
\details{
We will compare two glm models and plot the rates using model with tariff classes. The original model with sort the clustering in the original order.
Use rating_factors() to compare the output obtained from two or more glm objects.
}
\examples{
data(MTPL)
age_frequency <- fit_gam(data = MTPL, nclaims = nclaims, x = age_policyholder, exposure = exposure)
k<-addclustvar(age_frequency,MTPL)
model1<- glm(nclaims ~ freq_clusters, offset = log(exposure), family = "poisson", data = k)
model2 <- glm(nclaims ~ freq_clusters + age_policyholder, offset = log(exposure), family = "poisson", data = k)
glmplotcomp(model1,model2,k)
}
\author{
Seonghwa Min
}
