% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmpred.R
\name{glmpred}
\alias{glmpred}
\title{adding predictions from glms}
\usage{
glmpred(object1, object2, model)
}
\arguments{
\item{object1}{glm model 1 (typically null)}

\item{object2}{glam model 2 (typically alternative)}

\item{model}{data set with tariff classes as a variable}
}
\value{
data set with predictions
}
\description{
Add prediction from GLM to the data set
}
\details{
We want to add predictions from the glms and add the prediction colum for each glm model.
}
\examples{
data(MTPL)
age_frequency <- fit_gam(data = MTPL, nclaims = nclaims, x = age_policyholder, exposure = exposure)
k<-addclustvar(age_frequency,MTPL)
model1<- glm(nclaims ~ freq_clusters, offset = log(exposure), family = "poisson", data = k)
model2 <- glm(nclaims ~ freq_clusters + age_policyholder, offset = log(exposure), family = "poisson", data = k)
glmpred(model1,model2,k)
}
\author{
Seonghwa Min
}
