% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influential.R
\name{influential}
\alias{influential}
\title{Influential points for linear regression}
\usage{
influential(data, y)
}
\arguments{
\item{data}{data set, preferrably data.frame}

\item{y}{outcome variable}
}
\value{
index of the influential point
}
\description{
Looking for Influential points
}
\details{
We want to find the influential points for a certain outcome variable in a data set using linear regression.
}
\examples{
influential(MTPL,MTPL$nclaims)
}
\author{
Seonghwa Min
}
