% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minpval.R
\name{minpval}
\alias{minpval}
\title{P-values and Confidence interval for multivariate linear regression}
\usage{
minpval(data, y)
}
\arguments{
\item{data}{data set, preferrably data.frame}

\item{y}{outcome variable}
}
\value{
list of pvalues and confidence interval for each variable
}
\description{
P-values for each variable in multivariate linear regression with C.I.
}
\details{
We want to get the p-values for each variables in the multivariate linear regression.
}
\examples{
minpval(MTPL,MTPL$nclaims)
}
\author{
Seonghwa Min
}
