% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{normality}
\alias{normality}
\title{Normality Test for standardized residuals}
\usage{
normality(data, y)
}
\arguments{
\item{data}{data set, preferrably data.frame}

\item{y}{outcome variable}
}
\value{
QQ plot and shapiro test p-value
}
\description{
Normality Test
}
\details{
We are trying to assess the normality of standardized residuals using visual inspection from Q-Q plot and using Shapiro-Wilk normality
test.
}
\examples{
data(MTPL) 
normality(MTPL,MTPL$nclaims)
}
\author{
Seonghwa Min
}
