% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelimmult.R
\name{prelimmult}
\alias{prelimmult}
\title{Preliminary multivariate linear regression analysis}
\usage{
prelimmult(data, y)
}
\arguments{
\item{data}{data set, preferrably data.frame}

\item{y}{outcome variable}
}
\value{
summary of coefficients
}
\description{
Preliminary multivariate linear regression
}
\details{
We want to look at how other variable are related to the outcome variable y.
}
\examples{
data(MTPL)
prelimmult(MTPL,MTPL$nclaims)
}
\author{
Seonghwa Min
}
